+++ samples/file.ex
@@ -1,3 +1,3 @@
-unless true
-else
-end
+IO.puts("hello world");
+IO.puts( 1+1 );
+