+++ samples/lib/samples.ex
@@ -1,0 +1,18 @@
+defmodule Samples do
+  @moduledoc """
+  Documentation for `Samples`.
+  """
+
+  @doc """
+  Hello world.
+
+  ## Examples
+
+      iex> Samples.hello()
+      :world
+
+  """
+  def hello do
+    :world
+  end
+end