+++ samples/test/samples_test.exs
@@ -1,0 +1,8 @@
+defmodule SamplesTest do
+  use ExUnit.Case
+  doctest Samples
+
+  test "greets the world" do
+    assert Samples.hello() == :world
+  end
+end