+++ src/linters/swiftlint.ts
@@ -62,7 +62,9 @@ export const getIgnoreFilePragma: LinterGetIgnoreFilePragmaFunction = ({
   let existingRules: string[] = [];
 
   if (matches) {
-    existingRules = (matches[1] ?? "").split(",").map((item) => item.trim());
+    existingRules.push(
+      ...(matches[1] ?? "").split(",").map((item) => item.trim()),
+    );
   }
 
   existingRules.push(code);