+++ samples/file.dart
@@ -1,0 +1,9 @@
+class whyOhWhy {}
+
+void explicitDynamicType(dynamic object) {
+  print(object.foo());
+
+  if (1 > 2) {
+    print("1");
+  } else {}
+}